Put some evaluation methods in seperate file
